fix min invocation 5

+++ src/competitors/gpu/gpu_convert.cu
@@ -74,10 +74,10 @@ namespace Competitors {
 		int nnz_per_thread = 64;
 
 		// set the number of threads per block
-		int threads_per_block = min(max_threads_per_block, 512);
+		int threads_per_block = min(static_cast<int>(max_threads_per_block), 512);
 
 		int max_num_threads = num_sm * max_threads_per_sm;
-		int num_threads = min((sparse_size + nnz_per_thread - 1) / nnz_per_thread, max_num_threads);
+		int num_threads = min((static_cast<int>((sparse_size + nnz_per_thread - 1) / nnz_per_thread), static_cast<int>(max_num_threads));
 		int num_thread_blocks = (num_threads + threads_per_block - 1) / threads_per_block;
 
 		// Convert to COO
